Fix build --install failure in SDK top level
When running build.py --install from top level, do not add clashing packaging/install rules.
The original if condition here only applied when --package was selected,
not for --install.
What we want is just to check if we are running from component level,
and only in that case bother with the packaging rules.

Change-Id: I7865c2bef086ef91d107c2d93681bf894a2b183c
Signed-off-by: Gergely Nagy <gergely.nagy@arm.com>