Forward errors/messages from SPIRV-Tools
Previously, if there was an error parsing a SPIRV module in isGraphSpirv
or tryGetExtInstVersion then no information would be available to the
user to help them diagnose the cause. This patch registers a callback
with the spvtools::BuildModule function which forwards the messages
from SPIRV-Tools to the emulation layer's logging system.

Change-Id: I816da5a98263173436d846754c735f29d2c917ac
Signed-off-by: Rob Hughes <robert.hughes@arm.com>